Make applying voucher methods more general
