+++ saleor/discount/forms.py
@@ -4,8 +4,8 @@ from django import forms
 from django.utils.encoding import smart_text
 from django.utils.translation import pgettext_lazy
 
+from ..checkout.utils import get_voucher_discount_for_checkout
 from .models import NotApplicable, Voucher
-from .utils import get_voucher_discount_for_checkout
 
 
 class VoucherField(forms.ModelChoiceField):
